+++ fwArduino/speedController.h
@@ -1,6 +1,6 @@
 // Chalkbot 2018
 
-#include "inttypes.h" // now types like uint32_t can be used.
+#include "inttypes.h" // now types like unsigned long can be used.
 
 // Speed Controller:
 // This class manages the state of the Chalk bot in terms of 
@@ -12,11 +12,11 @@ class speedController
     private:
 
     const unsigned int m_baseClk_us;
-    uint32_t m_interval;       // [ticks]
-    uint32_t m_targetInterval; // [ticks] 1/target speed
-    uint32_t m_acc;
-    uint32_t m_steps;    
-    uint32_t m_cumTicks;
+    unsigned long m_interval;       // [ticks]
+    unsigned long m_targetInterval; // [ticks] 1/target speed
+    unsigned long m_acc;
+    unsigned long m_steps;    
+    unsigned long m_cumTicks;
     bool m_first;        // perform first step after go() at next call of isNextStep()
  
     public:
@@ -25,11 +25,11 @@ class speedController
 
     // Returns the time to the next Step in ticks
     // ticks - number of ticks since last step todo handle overflow
-    int32_t timeToNextStep( uint32_t ticks ) {};
+    long timeToNextStep( unsigned long ticks ) {};
 
     // Returns if next step is due
     // ticks - number of ticks since last step has been performed todo handle overflow
-    bool isNextStep(uint32_t ticks );
+    bool isNextStep(unsigned long ticks );
 
     // Inform class that step has been made
     void step() { m_steps--; };
@@ -43,7 +43,7 @@ class speedController
 
     // Set maximum acceleration that is usually the constant acceleration 
     // used while ramp up and ramp down
-    void setMaxAcceleration( uint32_t maxAcceleration ) { m_acc = maxAcceleration; };
+    void setMaxAcceleration( unsigned long maxAcceleration ) { m_acc = maxAcceleration; };
 
     // Start the Motion with ramp up
     // steps - number of steps to go
@@ -52,7 +52,7 @@ class speedController
     //         limited by maximum accelaration and number of steps if the ramp up can not
     //         be completed before the ramp down has to start.
     //         Allowed range 1 to 1000
-    uint32_t go( uint32_t steps , uint32_t speed );
+    unsigned long go( unsigned long steps , unsigned long speed );
 
     // Stop immediately
     void stop() { m_steps = 0; };
